reduce tags more
